requirements.txt
ultralytics>=8.3,<9
PyYAML>=6.0
numpy>=1.23
opencv-python>=4.8
